% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_lv_table.R
\name{cor_lv_table}
\alias{cor_lv_table}
\title{Model-implied correlation table with significance asterisks}
\usage{
cor_lv_table(fit)
}
\arguments{
\item{fit}{}
}
\description{
\code{cor_lv_table} takes a 'lavaan' object and returns a correlation table with significance asterisks
}
