% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esem2ewc.R
\name{esem2ewc}
\alias{esem2ewc}
\title{ESEM-within-CFA syntax generator}
\usage{
esem2ewc(esem.fit, referents)
}
\arguments{
\item{referents}{}
}
\description{
\code{esem2ewc} takes an ESEM fit from lavaan and generates the corresponding ESEM-within-CFA syntax. (for each latent variable, one referent item must be supplied.)
}
