% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esem_loadings.R
\name{esem_loadings}
\alias{esem_loadings}
\title{ESEM loading pattern matrix}
\usage{
esem_loadings(fit, digits = 3)
}
\arguments{
\item{digits}{}
}
\description{
\code{esem_loadings} takes an ESEM fit from lavaan and returns a loading pattern matrix. Currently, only traditional ESEM models are supported, not other extended methods such as set ESEM.
}
