% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esem_syntax.R
\name{esem_syntax}
\alias{esem_syntax}
\title{Generate ESEM model syntax automatically}
\usage{
esem_syntax(efa.fit, anchors)
}
\arguments{
\item{efa.fit}{}

\item{anchors}{}
}
\description{
This function generates an ESEM model syntax based on a previously fitted EFA model and a list of anchor variables or items
}
