% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_target.R
\name{fa.target}
\alias{fa.target}
\title{Fit an EFA with target rotation}
\usage{
fa.target(model, data, n_iter = NULL)
}
\description{
This function fits an EFA with an oblique target rotation, with the target keys extracted from a lavaan model syntax
}
