% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_table.R
\name{fit.table}
\alias{fit.table}
\title{Extract CFI, SRMR, and RMSEA into a table}
\usage{
fit.table(fit)
}
\description{
This function extracts CFI, SRMR, and RMSEA from a previously fitted lavaan model into a table syntax that can be paste into a word document and then be converted into a table (only robust measures sofar)
}
