% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_text.R
\name{fit.text}
\alias{fit.text}
\title{Extract CFI, SRMR, and RMSEA into a text report}
\usage{
fit.text(fit)
}
\description{
This function CFI, SRMR, and RMSEA from a previously fitted lavaan model into a text report (only robust measures sofar)
}
