% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_heat.R
\name{heat_resid}
\alias{heat_resid}
\title{Heat map for residuals}
\usage{
heat_resid(fit, threshold = FALSE, show.values = FALSE, font.size = NA)
}
\arguments{
\item{font.size}{}
}
\description{
\code{resid_heat} return a heat map, depicting the residual covariance among observed variables
}
