% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_syntax.R
\name{med_syntax}
\alias{med_syntax}
\title{Generate mediation analysis syntax for lavaan
\Code{med_syntax} generates mediation analysis syntax from a CFA model, with the first latent variable as the independent, the last as dependent, and the rest as mediators.}
\usage{
med_syntax(fit.cfa)
}
\arguments{
\item{fit.cfa}{}
}
\description{
Generate mediation analysis syntax for lavaan
\Code{med_syntax} generates mediation analysis syntax from a CFA model, with the first latent variable as the independent, the last as dependent, and the rest as mediators.
}
