% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel_it.R
\name{parcel_it}
\alias{parcel_it}
\title{Generate parceled data set}
\usage{
parcel_it(data, model, p.size = 2)
}
\description{
This function Creates a data frame with the parcel scores calculated for items within subscales based on a CFA structure
}
