% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_referent.R
\name{pick_referent}
\alias{pick_referent}
\title{Pick referent variables to be used in ESEM-within-CFA}
\usage{
pick_referent(esem.fit)
}
\arguments{
\item{esem.fit}{A traditional ESEM fitted model of class \code{lavaan}}
}
\value{
A named list containing one referent variable per each latent factor
}
\description{
\code{pick_referent} takes an ESEM fit and select referent variables to be used in ESEM-within-CFA.
}
\examples{
# fit the CFA model

cfa.mod <- 'visual  =~ x1 + x2 + x3
            textual =~ x4 + x5 + x6
            speed   =~ x7 + x8 + x9'

cfa.fit <- lavaan::cfa(cfa.mod, data = HolzingerSwineford1939, estimator = 'MLR')

# fit the ESEM model

esem.mod <- esemTools::cfa2esem(cfa.fit)

esem.fit <- lavaan::cfa(esem.mod, data = HolzingerSwineford1939, estimator = 'MLR')

# select referent variables

refs <- pick_referent(esem.fit)

# fit the EWC

ewc.mod <- esemTools::esem2ewc(esem.fit, referents = refs)

ewc.fit <- lavaan::cfa(ewc.mod, data = HolzingerSwineford1939, estimator = 'MLR')

}
