% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{R2}
\alias{R2}
\title{R-Squared from Correlation Matrix}
\usage{
R2(r)
}
\description{
This function takes a correlation matrix and calculates multiple squared correlation between each variable and all other variables
}
