% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_scores.R
\name{sum_scores}
\alias{sum_scores}
\title{Calculate Sum Scores from CFA Model}
\usage{
sum_scores(data, model)
}
\arguments{
\item{data}{}

\item{model}{}
}
\description{
This function takes a data frame and a CFA model syntax and calculates sum scores for each latent variable
}
