% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_key.R
\name{tar.key}
\alias{tar.key}
\title{Make target key for target rotation}
\usage{
tar.key(model, data)
}
\arguments{
\item{data}{}
}
\description{
\code{target_key} takes a CFA fit from lavaan and creates a target key to be used in ESEM or EFA with target rotation.
}
