% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa2esem.R
\name{cfa2esem}
\alias{cfa2esem}
\title{Generate ESEM syntax from a CFA fit}
\usage{
cfa2esem(cfa.fit)
}
\arguments{
\item{cfa.fit}{}
}
\description{
\code{cfa2esem} takes a CFA fit from lavaan and returns its corresponding ESEM syntax.
}
